Remove Component folder and dispatch its features
