+++ app/conf/config.json.template
@@ -1,11 +1,11 @@
 {
-    "file": "CoreBot configuation",
+    "file": "CoreBot configuration",
     "author": "guillain",
     "license": "LGPL-3.0",
     "description": [
         "use this template to create your own chatbot scenario",
-        "nothing more is need if you dont need specific feature",
-        "For new feature, read the doc"
+        "nothing more is need if you don't need specific feature",
+        "and for new feature, read the appropriate doc"
     ],
     "log": {
         "debug": true,
@@ -120,7 +120,7 @@
         }
     },
     "module": {
-        "autoreply": {
+        "auto-reply": {
             "enable": false,
             "delay": 1,
             "file": "data/autoreply.csv"
