+++ app/controller/loader.js
@@ -22,14 +22,14 @@ exports.run = function(controller, config) {
     _.each(config.controller.hears, function (conf, index) {
         if (conf_merged.controller.hears[index].enable === true) {
             tools.debug("info","controller loader hears " + index);
-            let mod_run = require(__basedir + 'controller/hears/' + index + '/run.js');
+            let mod_run = require(__basedir + 'controller/hears/' + index + '/stanza.js');
             controller = mod_run.run(controller, conf_merged);
         } else tools.debug("debug","not controller loader hears " + index);
     });
     _.each(config.controller.on, function (conf, index) {
         if (conf_merged.controller.on[index].enable === true) {
             tools.debug("debug","controller loader on " + index);
-            let mod_run = require(__basedir + 'controller/on/' + index + '/run.js');
+            let mod_run = require(__basedir + 'controller/on/' + index + '/stanza.js');
             controller = mod_run.run(controller, conf_merged);
         } else tools.debug("debug","not controller loader on " + index);
     });
