+++ app/module/auto-reply/run.js
@@ -2,15 +2,15 @@
 let tools = require(__basedir + 'lib/tools');
 
 // Requirements
-let csv = require(__basedir + 'module/csv/run.js');
+let csv = require(__basedir + 'module/csv/stanza.js');
 
 // AutoReply loader
 exports.run = function(bot, message, config){
-    tools.debug('debug', 'module autoreply run');
+    tools.debug('debug', 'module auto-reply run');
 
     // Reject bot message
     if (message.user === config.user) {
-        tools.debug("info", 'module autoreply run stop_bot_reply');
+        tools.debug("info", 'module auto-reply run stop_bot_reply');
         return;
     }
 
@@ -25,21 +25,21 @@ exports.run = function(bot, message, config){
 
     // Get CSV file data
     csv.get_csv_data_cb(bot, message, __basedir + config.module.autoreply.file, config, function(csv_data) {
-        tools.debug("info", 'module autoreply run csv_data  len ' + csv_data.length);
+        tools.debug("info", 'module auto-reply run csv_data  len ' + csv_data.length);
 
         // Pickup one entry
         let index = Math.floor(Math.random() * csv_data.length);
         let csv_picked_up = csv_data[index];
-        tools.debug('info', ' module autoreply run csv_picked_up index ' + index + ' csv_picked_up ' + csv_picked_up[0]);
+        tools.debug('info', ' module auto-reply run csv_picked_up index ' + index + ' csv_picked_up ' + csv_picked_up[0]);
 
         // Delay
-        tools.debug('info', 'module autoreply run delay ' + config.module.autoreply.delay*1000);
+        tools.debug('info', 'module auto-reply run delay ' + config.module.autoreply.delay*1000);
         setTimeout(function () {
             // Send reply
             let msg_to_send = csv_picked_up[0] + '\n' + csv_picked_up[2];
             //let msg_to_send = csv_picked_up[0] + '\n' + csv_picked_up[1] + '\n' + csv_picked_up[2];
             //let msg_to_send = '<a href="' + csv_picked_up[2] + '">' + csv_picked_up[0] + '</a>';
-            tools.debug('debug', 'module autoreply run send_reply ' + msg_to_send);
+            tools.debug('debug', 'module auto-reply run send_reply ' + msg_to_send);
             bot.reply(message, msg_to_send);
         }, config.module.autoreply.delay * 1000);
     });
