+++ app/module/search/conf.json
@@ -4,7 +4,6 @@
             "enable": false,
             "prefixed": true,
             "limit": 10,
-            "sub_limit": 2,
             "key_length": 2,
             "storage": "",
             "msg": {
