+++ app/module/service-desk/conf.json
@@ -1,0 +1,18 @@
+{
+    "module": {
+        "service-desk": {
+            "enable": false,
+            "prefixed": true,
+            "storage": "",
+            "msg": {
+                "help": [
+                    "Service desk is a support for the employees\n",
+                    "- _No argument_\n",
+                    "_Example_: `hello`",
+                    "Dialog with the service desk bot to reach your expectation"
+                ]
+            }
+        }
+    }
+}
+
