+++ doc/controller.md
@@ -19,7 +19,7 @@ loaded or not.
 They are define in dedicated folder and they need the following files:
 - Controller name
   - hears or on
-    - *run.js*: Scripts with the code
+    - *stanza.js*: Scripts with the code
     - *conf.js*: Default configuration file
 
 They are located in the './controller' folder.
