+++ doc/launcher.md
@@ -20,7 +20,7 @@ the Controller loaded.
 They are define in dedicated folder and they need at least the
 following files:
 - Module name
-  - *run.js*: Scripts with the code
+  - *stanza.js*: Scripts with the code
   - *conf.js*: Default configuration file
 
 They are located in the './module' folder.
